+++ src/frontend/src/eth/services/erc20.services.ts
@@ -26,7 +26,6 @@ import { loadNetworkCustomTokens } from '$lib/services/custom-tokens.services';
 import { i18n } from '$lib/stores/i18n.store';
 import { toastsError, toastsErrorNoTrace } from '$lib/stores/toasts.store';
 import type { LoadCustomTokenParams } from '$lib/types/custom-token';
-import type { OptionIdentity } from '$lib/types/identity';
 import type { NetworkId } from '$lib/types/network';
 import type { ResultSuccess } from '$lib/types/utils';
 import { parseCustomTokenId } from '$lib/utils/custom-token.utils';
@@ -76,22 +75,23 @@ const loadDefaultErc20Tokens = async (): Promise<ResultSuccess> => {
 	return { success: true };
 };
 
-export const loadCustomTokens = ({
+export const loadCustomTokens = async ({
 	identity,
 	useCache = false,
 	tokens,
-	certified: restCertified
-}: Omit<LoadCustomTokenParams, 'certified'> & {
-	certified?: boolean;
-	tokens?: CustomToken[];
-}): Promise<void> => {
+	certified
+}: LoadCustomTokenParams): Promise<void> => {
 	if (nonNullish(tokens)) {
-		return loadCustomTokensWithMetadata({
+		const response = await loadCustomTokensWithMetadata({
 			identity,
-			certified: restCertified ?? true,
+			certified,
 			useCache,
 			tokens
-		}).then((response) => loadCustomTokenData({ response, certified: restCertified ?? true }));
+		});
+
+		loadCustomTokenData({ response, certified });
+
+		return;
 	}
 
 	return queryAndUpdate<Erc20CustomToken[]>({
