+++ src/frontend/src/lib/services/custom-tokens.services.ts
@@ -151,86 +151,3 @@ export const loadNetworkCustomTokens = async ({
 		filterTokens
 	});
 };
-
-export const loadCustomTokens = async ({
-	identity,
-	...rest
-}: {
-	identity: OptionIdentity;
-} & Omit<LoadCustomTokensParams, 'identity' | 'filterTokens' | 'certified'>): Promise<void> => {
-	if (isNullish(identity)) {
-		return;
-	}
-
-	await queryAndUpdate<CustomToken[]>({
-		request: ({ certified }) =>
-			loadNetworkCustomTokens({
-				identity,
-				certified,
-				...rest
-			}),
-		onLoad: async ({ response: tokens, certified }) => {
-			const { loadCustomTokens: loadIcrcCustomTokens } = await import(
-				'../../icp/services/icrc.services'
-			);
-			const { loadCustomTokens: loadErc20CustomTokens } = await import(
-				'../../eth/services/erc20.services'
-			);
-			const { loadCustomTokens: loadErc721CustomTokens } = await import(
-				'../../eth/services/erc721.services'
-			);
-			const { loadCustomTokens: loadErc1155CustomTokens } = await import(
-				'../../eth/services/erc1155.services'
-			);
-			const { loadCustomTokens: loadSplCustomTokens } = await import(
-				'../../sol/services/spl.services'
-			);
-			const { loadCustomTokens: loadExtCustomTokens } = await import(
-				'../../icp/services/ext.services'
-			);
-			const { loadCustomTokens: loadIcPunksCustomTokens } = await import(
-				'../../icp/services/icpunks.services'
-			);
-
-			const icrcTokens: CustomToken[] = [];
-			const erc20Tokens: CustomToken[] = [];
-			const erc721Tokens: CustomToken[] = [];
-			const erc1155Tokens: CustomToken[] = [];
-			const splTokens: CustomToken[] = [];
-			const extTokens: CustomToken[] = [];
-			const icPunksTokens: CustomToken[] = [];
-
-			tokens.forEach((token) => {
-				if ('Icrc' in token.token) {
-					icrcTokens.push(token);
-				} else if ('Erc20' in token.token) {
-					erc20Tokens.push(token);
-				} else if ('Erc721' in token.token) {
-					erc721Tokens.push(token);
-				} else if ('Erc1155' in token.token) {
-					erc1155Tokens.push(token);
-				} else if ('SplMainnet' in token.token || 'SplDevnet' in token.token) {
-					splTokens.push(token);
-				} else if ('ExtV2' in token.token) {
-					extTokens.push(token);
-				} else if ('IcPunks' in token.token) {
-					icPunksTokens.push(token);
-				}
-			});
-
-			await Promise.all([
-				loadIcrcCustomTokens({ identity, certified, tokens: icrcTokens }),
-				loadErc20CustomTokens({ identity, certified, tokens: erc20Tokens }),
-				loadErc721CustomTokens({ identity, certified, tokens: erc721Tokens }),
-				loadErc1155CustomTokens({ identity, certified, tokens: erc1155Tokens }),
-				loadSplCustomTokens({ identity, certified, tokens: splTokens }),
-				loadExtCustomTokens({ identity, certified, tokens: extTokens }),
-				loadIcPunksCustomTokens({ identity, certified, tokens: icPunksTokens })
-			]);
-		},
-		onUpdateError: () => {
-			// Errors are handled standard by standard
-		},
-		identity
-	});
-};
