+++ src/frontend/src/lib/types/custom-token.ts
@@ -1,4 +1,4 @@
-import type { Token as BackendToken } from '$declarations/backend/backend.did';
+import type { CustomToken as BackendCustomToken, Token as BackendToken } from '$declarations/backend/backend.did';
 import type { Erc20Token } from '$eth/types/erc20';
 import type { Dip721Token } from '$icp/types/dip721-token';
 import type { ExtToken } from '$icp/types/ext-token';
@@ -11,8 +11,8 @@ import type { QueryAndUpdateRequestParams } from '@dfinity/utils';
 
 type CustomTokenNetworkKeys = BackendToken extends infer T
 	? T extends Record<string, unknown>
-		? keyof T
-		: never
+	? keyof T
+	: never
 	: never;
 
 export type TokenVariant<K extends CustomTokenNetworkKeys, T> = T & { networkKey: K };
@@ -59,4 +59,8 @@ export type SaveCustomExtVariant = CustomTokenState & TokenVariant<'ExtV2', ExtS
 
 export type CustomToken<T extends Token> = TokenToggleable<T>;
 
-export type LoadCustomTokenParams = QueryAndUpdateRequestParams & { useCache?: boolean };
+export type LoadCustomTokenParams = QueryAndUpdateRequestParams & {
+	useCache?: boolean;
+	tokens?: BackendCustomToken[];
+	certified?: boolean;
+};
